Refactor the code for fixing the search/replace
